Fix controls don't appear when changing tabs.
Github seems to use some kind of single page application architecture, so the
extension was not being initialized as the user moved between tabs on the
PR page.

I fixed this by monitoring the DOMSubtreeModified event on the body, and
injecting the controls whenever they were not present but the sidebar was.